requirements.txt
llama-index
llama-index-vector-stores-chroma
chromadb
streamlit
python-dotenv
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-parse
pypdf
